Replace README.md with an archived project notice
The functionality that this DLL provided has been incorporated into
https://github.com/0xC0000054/sc4-resource-loading-hooks.

The sc4-resource-loading-hooks project was created to provide other DLL
plugins for SimCity 4 with the ability to register for notifications
when the game loads a resource that they want to log or modify.

SimCity 4 only allows one DLL to hook the loading code for a specific
resource type, with the first DLL to do so winning.
Because of this there is no guarantee which DLL would win if both are
installed.